kernel and mod0: Calling the module's entry point.
kernel-mod.c:
Implementing the caller routine to call the entrypoint in mod0.bin.
Working on the inline assembly to pass two parameters to the function.
This is a work in progress.

mod0.bin and mod1.bin:
Changing the prototype for the crt0 entry point.
Now we need to receive two parameters.
This is a work in progress.

ps:
The parameters are 64bit long now.